+++ CHANGELOG.md
@@ -6,6 +6,8 @@ with Flutter's [Color] class. They were on a `0` to `1` scale.
 * Renamed [HsvColor] to [HsbColor] to avoid a naming conflict with Flutter's
 [Color] class's [value] parameter.
 
+* Renamed [interpolateTo] to [lerpTo].
+
 ## [0.2.8] - March 29, 2020
 
 * Removed [isBlack] and [isWhite] conversion condition from the CMYK color space.
@@ -17,7 +19,7 @@ RGB values equaling `255`.
 
 ## [0.2.7] - March 29, 2020
 
-* Added the [lerpTo] method to each color model.
+* Added the [interpolateTo] method to each color model.
 
 * Override the conversion methods on each color model that return their own
 respective color spaces. Colors were being unnecessarily converted back and